requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1
bcrypt==4.2.0
email-validator==2.2.0
pytest==8.3.3
